(maint) Fix tests for Puppet 6+
For Puppet 6.0.1, some changes were made that affect Bolt. The way we
currently use bolt_spec doesn't use Bolt's vendored Puppet, it uses the
gem. So we need to manually keep compatible Puppet and Bolt versions in
sync. Pin to Bolt 0.23 and Puppet 6.0.0 until the next Bolt and Puppet
releases.