(BOLT-915) Make bash.sh facts implementation shareable
The puppetlabs-puppet_agent install_shell.sh implementation relies on determining target platform and version. With the introduction of multi-file task support this commit facilitates the use of the facts::bash implementation. The bash.sh implementation can now accept a positional argument that can be set to `platform` or `version`. This results in the requested data being printed to stdout instead of the full JSON report that would be printed without the positional argument.