(BOLT-997) Only use facter when puppet-agent feature is available
Previously the powershell and bash implementations would delegate to facter if it was available. This led to discrepencies in results based on the presense of puppet-agent on the target. Now the ruby implementation which requires the puppet-agent feature is the only implementation that delegates to facter.

The ruby task will also handle the case where legacy facter is installed on a target by not invoking facter with the --show-legacy flag.