(BOLT-997) Clean bash implementation for RedHat
When `lsb_release` is used to find platform name in the bash implementation for some red-hat versions (example redhat-5-x86_64) the OS name fact reports `RedHatEnterpriseServer`. This commit changes shortens the output to `RedHat` as the puppet_agent::install task expects.

When a target is missing both `lsb_release` and `os-release` pick up the name and version from `redhat-release`.