Source facts for systemd platforms using bash 3.1 and 3.2
This fixes a bug where facts were not sourced correctly from
`/etc/os-release` or `/usr/lib/os-release` on systemd platforms when
using bash 3.1 or 3.2. Previously, the `os-release` file was sourced
by using process substitution, which is incompatible with some versions
of bash. This updates the script to source from `/dev/stdin` instead.