Delegate to facter if installed
This restores the behavior that was removed in 9b7fe39, whereby the
shell and powershell implementations of the facts task will detect
whether `facter` is available and call out to it if possible.

This was removed because it leads to a situation where you can get
different kinds of output from the same implementation of the task
depending on where it's being run. But since the unit of abstraction is
the task itself, and not an implementation, it's better to provide the
full `facter` output in as many cases as possible, without the user
having to explicitly mark the target as having the `puppet-agent`
feature first. This makes the output less consistent between targets for
the same implementation, but more consistent for the same target between
implementations.