(bug-fix) Use RbConfig.ruby instead of hardcoded path
We rely on the puppet-agent being installed for targets that have the
`pupet-agent` feature. However we assign `puppet-agent` to localhost
expecting it will use Bolt's puppet-agent, which means puppet resources
aren't at the expected `/opt/puppetlabs/puppet` path. This will now look
for facter packaged with Puppet first, then packaged with
Bolt, then lastly will just call `facter` to see if it's elsewhere on
the PATH.

This also makes a number of modifications to how the executable string
is munged and called on Windows, so that the task is runnable on that
platform. It also uses the ruby_task_helper module in order to catch and
display errors.