(GH-1981) Add external facts plan
This adds the prototype for a plan that takes the absolute path to a
directory containing external facts scripts. It runs each script and
munges the output to a Hash, then adds the facts to the Bolt Target
objects and returns the targets. This is useful for gathering
information from remote systems without relying on facter or a bash
script, and allows users to easily create their own facts and handle
them in Bolt plans.